clean architecture categories
