+++ lib/app.dart
@@ -3,11 +3,12 @@ import 'package:man_shop_app/config/themes/app_themes.dart';
 import 'package:man_shop_app/core/utils/app_strings.dart';
 import 'package:man_shop_app/data/repositories/authentication/logout_repository/logout_repository.dart';
 import 'package:man_shop_app/data/repositories/authentication/register_repository/register_repository.dart';
-import 'package:man_shop_app/data/repositories/home/home_repository.dart';
 import 'package:man_shop_app/data/web_service/authentication/login_web_service.dart';
 import 'package:man_shop_app/data/web_service/authentication/logout_web_service.dart';
 import 'package:man_shop_app/data/web_service/authentication/register_web_service.dart';
-import 'package:man_shop_app/data/web_service/home/home_web_service.dart';
+import 'package:man_shop_app/features/home/data/repositories/home_repository.dart';
+import 'package:man_shop_app/features/home/data/web_service/home_web_service.dart';
+import 'package:man_shop_app/features/home/presentation/bloc/home_cubit.dart';
 import 'package:man_shop_app/presentation/authentication/login/bloc/login_cubit.dart';
 import 'package:man_shop_app/presentation/authentication/login/screens/login_screen.dart';
 import 'package:man_shop_app/presentation/authentication/logout/bloc/logout_cubit.dart';
@@ -15,7 +16,6 @@ import 'package:man_shop_app/presentation/authentication/register/bloc/register_
 import 'package:man_shop_app/presentation/bottom_navigation_bar/bloc/cubit.dart';
 import 'package:man_shop_app/presentation/bottom_navigation_bar/screens/Bottom_nav_bar_screen.dart';
 import 'package:man_shop_app/presentation/on_boarding/screens/on_boarding_screen.dart';
-import 'package:man_shop_app/presentation/home/bloc/home_cubit.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'data/repositories/authentication/login_repository/login_repository.dart';
