--- lib/data/models/home/categories_model.dart
@@ -1,85 +1,0 @@
-class CategoriesModel {
-  CategoriesModel({
-    required this.currentPage,
-    required this.data,
-    required this.firstPageUrl,
-    required this.from,
-    required this.lastPage,
-    required this.lastPageUrl,
-    //this.nextPageUrl,
-    required this.path,
-    required this.perPage,
-    //this.prevPageUrl,
-    required this.to,
-    required this.total,
-  });
-  late final int currentPage;
-  late final List<Data> data;
-  late final String firstPageUrl;
-  late final int from;
-  late final int lastPage;
-  late final String lastPageUrl;
-  //late final Null nextPageUrl;
-  late final String path;
-  late final int perPage;
-  //late final Null prevPageUrl;
-  late final int to;
-  late final int total;
-
-  CategoriesModel.fromJson(Map<String, dynamic> json){
-    currentPage = json['current_page'];
-    data = List.from(json['data']).map((e)=>Data.fromJson(e)).toList();
-    firstPageUrl = json['first_page_url'];
-    from = json['from'];
-    lastPage = json['last_page'];
-    lastPageUrl = json['last_page_url'];
-    //nextPageUrl = json['next_page_url'];
-    path = json['path'];
-    perPage = json['per_page'];
-    //prevPageUrl = json['prev_page_url'];
-    to = json['to'];
-    total = json['total'];
-  }
-
-  Map<String, dynamic> toJson() {
-    final _data = <String, dynamic>{};
-    _data['current_page'] = currentPage;
-    _data['data'] = data.map((e)=>e.toJson()).toList();
-    _data['first_page_url'] = firstPageUrl;
-    _data['from'] = from;
-    _data['last_page'] = lastPage;
-    _data['last_page_url'] = lastPageUrl;
-    //_data['next_page_url'] = nextPageUrl;
-    _data['path'] = path;
-    _data['per_page'] = perPage;
-    //_data['prev_page_url'] = prevPageUrl;
-    _data['to'] = to;
-    _data['total'] = total;
-    return _data;
-  }
-}
-
-class Data {
-  Data({
-    required this.id,
-    required this.name,
-    required this.image,
-  });
-  late final int id;
-  late final String name;
-  late final String image;
-
-  Data.fromJson(Map<String, dynamic> json){
-    id = json['id'];
-    name = json['name'];
-    image = json['image'];
-  }
-
-  Map<String, dynamic> toJson() {
-    final _data = <String, dynamic>{};
-    _data['id'] = id;
-    _data['name'] = name;
-    _data['image'] = image;
-    return _data;
-  }
-}
