+++ lib/features/home/data/repositories/home_repository.dart
@@ -1,5 +1,6 @@
-import 'package:man_shop_app/data/models/home/home_model.dart';
-import 'package:man_shop_app/data/web_service/home/home_web_service.dart';
+import 'package:man_shop_app/features/home/data/models/categories_model.dart';
+import 'package:man_shop_app/features/home/data/models/home_model.dart';
+import 'package:man_shop_app/features/home/data/web_service/home_web_service.dart';
 
 class HomeRepository{
   HomeWebService homeWebService;
@@ -8,5 +9,10 @@ class HomeRepository{
   Future<HomeModel> getHomeData() async {
     final data = await homeWebService.getHomeData();
     return HomeModel.fromJson(data);
+  }
+
+  Future<CategoriesModel> getHomeCategories() async {
+    final data = await homeWebService.getHomeCategories();
+    return CategoriesModel.fromJson(data['data']);
   }
 }
