+++ lib/features/home/presentation/bloc/home_cubit.dart
@@ -1,23 +1,25 @@
 import 'dart:async';
 import 'dart:developer';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:man_shop_app/data/models/home/home_model.dart';
-import 'package:man_shop_app/data/repositories/home/home_repository.dart';
-import 'package:man_shop_app/presentation/home/bloc/home_states.dart';
-import 'package:man_shop_app/shared/components/constants.dart';
-import 'package:man_shop_app/shared/network/end_points.dart';
-import 'package:man_shop_app/shared/network/remote/dio_helper.dart';
+import 'package:man_shop_app/features/home/data/models/categories_model.dart';
+import 'package:man_shop_app/features/home/data/models/home_model.dart';
+import 'package:man_shop_app/features/home/data/repositories/home_repository.dart';
+import 'package:man_shop_app/features/home/presentation/bloc/home_states.dart';
+
 
 class HomeCubit extends Cubit<HomeStates> {
   HomeCubit(this.homeRepository) : super(HomeInitialState());
 
   late HomeModel homeModel;
+  late CategoriesModel categoriesModel;
   late HomeRepository homeRepository;
+  int indicatorIndex = 0;
 
   Future<void> getHomeData() async {
     emit(HomeLoadingState());
     try {
       homeModel = await homeRepository.getHomeData();
+      categoriesModel = await homeRepository.getHomeCategories();
       emit(HomeSuccessState());
     } catch (error, s) {
       log(error.toString());
@@ -25,8 +27,6 @@ class HomeCubit extends Cubit<HomeStates> {
       emit(HomeErrorState(error.toString()));
     }
   }
-
-  int indicatorIndex = 0;
 
   void changeBannerIndex(index) {
     indicatorIndex = index;
