+++ lib/features/home/presentation/screens/home_screen.dart
@@ -2,10 +2,12 @@ import 'dart:developer';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:man_shop_app/presentation/Home/widgets/banner_widget.dart';
-import 'package:man_shop_app/presentation/home/bloc/home_cubit.dart';
-import 'package:man_shop_app/presentation/home/bloc/home_states.dart';
-import 'package:man_shop_app/presentation/home/widgets/products_widget.dart';
+import 'package:man_shop_app/features/home/presentation/bloc/home_cubit.dart';
+import 'package:man_shop_app/features/home/presentation/bloc/home_states.dart';
+import 'package:man_shop_app/features/home/presentation/widgets/banner_widget.dart';
+import 'package:man_shop_app/features/home/presentation/widgets/categories_bar.dart';
+import 'package:man_shop_app/features/home/presentation/widgets/products_widget.dart';
+
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({Key? key}) : super(key: key);
@@ -27,10 +29,24 @@ class HomeScreen extends StatelessWidget {
                   )
                 : SingleChildScrollView(
                   child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
                     children: [
                       const SizedBox(height: 85.0),
-                      BannerWidget(),
-                      const SizedBox(height: 15.0),
+                      const BannerWidget(),
+                      const SizedBox(height: 30.0),
+                      const CategoriesBar(),
+                      const SizedBox(height: 30.0),
+                      const Padding(
+                        padding: EdgeInsets.only(left: 15.0,bottom: 10.0),
+                        child: Text(
+                          'Products',
+                          style: TextStyle(
+                            fontSize: 20,
+                            fontWeight: FontWeight.bold,
+                          ),
+                          textAlign: TextAlign.left,
+                        ),
+                      ),
                       ProductsWidget(products: homeCubit.homeModel.data.products,),
                     ],
                   ),
