+++ lib/features/home/presentation/widgets/categories_bar.dart
@@ -1,0 +1,81 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:man_shop_app/features/home/presentation/bloc/home_cubit.dart';
+import 'package:man_shop_app/features/home/presentation/bloc/home_states.dart';
+
+class CategoriesBar extends StatelessWidget {
+  const CategoriesBar({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final homeCubit = BlocProvider.of<HomeCubit>(context);
+    return BlocConsumer<HomeCubit, HomeStates>(
+      listener: (context, state) {
+        // TODO: implement listener
+      },
+      builder: (context, state) {
+        return SizedBox(
+          width: double.infinity,
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              const Padding(
+                padding: EdgeInsets.only(left: 15.0,bottom: 10.0),
+                child: Text(
+                  'Categories',
+                  style: TextStyle(
+                    fontSize: 20,
+                    fontWeight: FontWeight.bold,
+                  ),
+                  textAlign: TextAlign.left,
+                ),
+              ),
+              Container(
+                color: Colors.grey[200],
+                padding: const EdgeInsets.symmetric(vertical: 10.0),
+                height: 145,
+                child: ListView.separated(
+                  shrinkWrap: true,
+                  scrollDirection: Axis.horizontal,
+                  itemCount: homeCubit.categoriesModel.data.length,
+                  padding: const EdgeInsets.only(left: 15.0, right: 15.0),
+                  itemBuilder: (context, index) {
+                    return Column(
+                      children: [
+                        SizedBox(
+                          width: 100,
+                          height: 100,
+                          child: ClipOval(
+                            child: Image.network(
+                              homeCubit.categoriesModel.data[index].image,
+                              fit: BoxFit.cover,
+                            ),
+                          ),
+                        ),
+                        Padding(
+                          padding: const EdgeInsets.only(top: 10.0),
+                          child: Text(
+                            homeCubit.categoriesModel.data[index].name,
+                            style: const TextStyle(
+                              fontSize: 12,
+                              //fontWeight: FontWeight.bold,
+                            ),
+                          ),
+                        ),
+                      ],
+                    );
+                  },
+                  separatorBuilder: (context, index) {
+                    return const SizedBox(
+                      width: 15,
+                    );
+                  },
+                ),
+              ),
+            ],
+          ),
+        );
+      },
+    );
+  }
+}
