+++ lib/features/home/presentation/widgets/products_widget.dart
@@ -1,14 +1,12 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_staggered_grid_view/flutter_staggered_grid_view.dart';
 import 'package:man_shop_app/core/utils/app_colors.dart';
-import 'package:man_shop_app/data/models/home/home_model.dart';
-import 'package:man_shop_app/shared/styles/color.dart';
+import 'package:man_shop_app/features/home/data/models/home_model.dart';
 
 class ProductsWidget extends StatelessWidget {
   final List<Products> products;
 
-  ProductsWidget({required this.products, Key? key}) : super(key: key);
+  const ProductsWidget({required this.products, Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -138,9 +136,9 @@ class ProductsWidget extends StatelessWidget {
                                     child: Container(
                                       height: 1.5,
                                       width: products[index]
-                                              .oldPrice
-                                              .toString()
-                                              .length *
+                                          .oldPrice
+                                          .toString()
+                                          .length *
                                           10,
                                       color: Colors.grey,
                                     ),
