--- lib/presentation/home/screens/home_screen.dart
@@ -1,43 +1,0 @@
-import 'dart:developer';
-
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:man_shop_app/presentation/Home/widgets/banner_widget.dart';
-import 'package:man_shop_app/presentation/home/bloc/home_cubit.dart';
-import 'package:man_shop_app/presentation/home/bloc/home_states.dart';
-import 'package:man_shop_app/presentation/home/widgets/products_widget.dart';
-
-class HomeScreen extends StatelessWidget {
-  const HomeScreen({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    final homeCubit = BlocProvider.of<HomeCubit>(context);
-    return BlocConsumer<HomeCubit, HomeStates>(
-      listener: (context, state) {},
-      builder: (context, state) {
-        return Scaffold(
-          body: SafeArea(
-            child: state is HomeLoadingState
-                ? const Center(
-                    child: CircularProgressIndicator(
-                      strokeWidth: 4,
-                      color: Colors.blue,
-                    ),
-                  )
-                : SingleChildScrollView(
-                  child: Column(
-                    children: [
-                      const SizedBox(height: 85.0),
-                      BannerWidget(),
-                      const SizedBox(height: 15.0),
-                      ProductsWidget(products: homeCubit.homeModel.data.products,),
-                    ],
-                  ),
-                ),
-          ),
-        );
-      },
-    );
-  }
-}
